Move SQLAlchemy NullPool suppression fixture to tests/unit/conftest.py
